fix(mobile): profile/settings/invite を SSR で native モード判定しハイドレーション不整合を解消 (#1037 round-3)
- src/app/(main)/profile/page.tsx, src/app/(main)/settings/page.tsx:
  Server Component 化して cookies() から is_native_app を読み、Client
  Component (ProfilePageClient/SettingsPageClient) へ initialIsNativeApp
  として渡すよう分割。SSR では Cookie 不在で isNativeApp が一旦 false に
  なり client hydration で true に切り替わることによる hydration
  mismatch + 破壊的ボタン(ログアウト/削除)の一瞬フラッシュを解消。
  (main)/MainLayout.tsx の BottomNav と同じ SSR-safe パターンを踏襲。
- src/app/invite/[token]/page.tsx:
  同様に Server Component 化し token と initialIsNativeApp を
  InvitePageClient へ渡す。「ログアウトしてやり直す」ボタンが
  isNativeApp 未ガードで signOut() を呼び、ネイティブと共有する
  refresh token を無効化しうる問題を解消(native時はsignOutを呼ばず
  ログイン画面へのみ遷移)。
- apps/mobile/app/settings/account.tsx:
  サーバー削除APIの失敗とローカルクリーンアップ(AsyncStorage/signOut)
  失敗でエラーメッセージを分離。前者は再試行可能である旨、後者は
  削除自体は完了済みである旨を案内する。
